Initial pass as using monotonic.monotonic instead of time.time
