+++ piksi_tools/console/tracking_view.py
@@ -10,9 +10,10 @@
 # EITHER EXPRESSED OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE IMPLIED
 # WARRANTIES OF MERCHANTABILITY AND/OR FITNESS FOR A PARTICULAR PURPOSE.
 
-import time
-from collections import defaultdict, deque
+import monotonic
 import threading
+
+from collections import defaultdict, deque
 from chaco.api import ArrayPlotData, Plot
 from chaco.tools.api import LegendTool
 from enable.api import ComponentEditor
@@ -142,8 +143,8 @@ class TrackingView(CodeFiltered):
     def measurement_state_callback(self, sbp_msg, **metadata):
         with self.CN0_lock:
             codes_that_came = []
-            t = time.time() - self.t_init
-            self.time.append(t)
+            the_time = monotonic.monotonic() - self.t_init
+            self.time.append(the_time)
             # first we loop over all the SIDs / channel keys we have stored and set 0 in for CN0
             for i, s in enumerate(sbp_msg.states):
                 if code_is_glo(s.mesid.code):
@@ -160,7 +161,7 @@ class TrackingView(CodeFiltered):
                 codes_that_came.append(key)
                 if s.cn0 != 0:
                     self.CN0_dict[key].append(s.cn0 / 4.0)
-                    self.CN0_age[key] = t
+                    self.CN0_age[key] = the_time
                 received_code_list = getattr(self, "received_codes", [])
                 if s.mesid.code not in received_code_list:
                     received_code_list.append(s.mesid.code)
@@ -168,13 +169,13 @@ class TrackingView(CodeFiltered):
             for key, cno_array in list(self.CN0_dict.items()):
                 if key not in codes_that_came:
                     cno_array.append(0)
-            self.clean_cn0(t)
+            self.clean_cn0(the_time)
         self.update_scheduler.schedule_update('update_plot', self.update_plot)
 
     def tracking_state_callback(self, sbp_msg, **metadata):
         with self.CN0_lock:
             codes_that_came = []
-            t = time.time() - self.t_init
+            t = monotonic.monotonic() - self.t_init
             self.time.append(t)
             # first we loop over all the SIDs / channel keys we have stored and set 0 in for CN0
             # for each SID, an array of size MAX PLOT with the history of CN0's stored
@@ -257,7 +258,7 @@ class TrackingView(CodeFiltered):
 
     def __init__(self, link):
         super(TrackingView, self).__init__()
-        self.t_init = time.time()
+        self.t_init = monotonic.monotonic()
         self.time = deque([x * 1 / TRK_RATE for x in range(-NUM_POINTS, 0, 1)], maxlen=NUM_POINTS)
         self.CN0_lock = threading.Lock()
         self.CN0_dict = defaultdict(lambda: deque([0] * NUM_POINTS, maxlen=NUM_POINTS))
