+++ requirements.txt
@@ -14,3 +14,4 @@ ruamel.yaml==0.15.87
 setuptools_scm==3.1.0
 enum34==1.1.6
 libsettings==0.1.11
+monotonic==1.5
